ASC-993 linter not respecting valid python styling
As a side effect of this work, we discovered that there
was a case where a validator could be supplied for a mark
when no values were supplied. This should never be allowed.

This commit adds logic to fail in that case. This new check
has an additional implication. Since we drop any non-string
value, we'll end up in a situation where it seems like no
value has been supplied, so we'll fail for both having a
non-string value and also for not having a value.